Add files via upload
The Bank Management System is a comprehensive software solution designed to manage and streamline the core operations of a bank. The system aims to enhance efficiency, ensure data accuracy, and provide a user-friendly interface for  customers. This project leverages the robust capabilities of Java for the backend and MySQL for the database, ensuring a reliable and scalable system.